requirements.txt
AI_ASSISTANT_FOR_HR==3.12.10
altair==5.5.0
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
embeddings==0.0.8
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
-e git+https://github.com/monamete/AI_ASSISTANT_FOR_HR.git@85923b999402cd3d6a2091c61301f13077ae8335#egg=mlproject
more-itertools==10.3.0
mpmath==1.3.0
narwhals==1.36.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.2.2
protobuf==5.29.4
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==79.0.1
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typeguard==4.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wheel==0.45.1
xgboost
zipp==3.19.2
